lake-flow/requirements.txt
# =========================
# Core runtime
# =========================
python-dotenv>=1.0.0

# =========================
# Data & Storage
# =========================
pandas>=2.0.0
numpy==1.26.4

# =========================
# File & Text processing
# =========================
PyPDF2>=3.0.0
pdfplumber>=0.11.0
pypdfium2<5.0
python-docx>=1.1.0
openpyxl>=3.1.0

# =========================
# AI / NLP (embedding)
# =========================
# Docker: install torch CPU first (Dockerfile). Mac M1 (venv): pip install torch then pip install -r requirements.txt → use GPU Metal (MPS).
sentence-transformers>=2.2.2
transformers>=4.36.0

# =========================
# Backend API (ready to use)
# =========================
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# =========================
# Utils
# =========================
tqdm>=4.66.0
loguru>=0.7.2

qdrant-client==1.16.2

# =========================
# Security / Auth
# =========================
passlib[bcrypt]>=1.7.4,<1.8
bcrypt<4.0
python-jose>=3.3.0

# =========================
# Dev / Test (API client, scripts)
# =========================
requests>=2.31.0

grpcio>=1.59.0,<2.0



